Fix modifier keys aren't shown in pressed state when two combination keys pressed consecutively. Fixes Issue 65.
Introduced a really_pressed property to TwoStateImage, which is used to record
if the key is really physically pressed or not. It is helpful to determine if
the key image should show unpressed key when the countdown times out. When
times out and the key is still really_pressed, then do not change back to
default image (unpressed). It will be changed when real key release event
comes
in.